django/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.6
django-bitfield==2.2.0
django-geo==0.8
django-heroku==0.3.1
django-js-asset==2.2.0
django-leaflet==0.30.0
django-mptt==0.16.0
django-separatedvaluesfield==0.6.0
django-timezone-field==6.1.0
django-uuidfield==0.5.0
djangorestframework==3.15.1
docopt==0.6.2
geographiclib==2.0
geopy==2.4.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
Js2Py==0.74
lxml==5.2.2
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
pipwin==0.5.2
psycopg2==2.9.9
pyjsparser==2.7.1
pypng==0.20220715.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
qrcode==7.4.2
requests==2.32.2
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
South==1.0.2
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
unicodecsv==0.14.1
urllib3==2.2.1
whitenoise==6.7.0
wsproto==1.2.0
